Cleanup Inside Out I/O Tutorial
`docs/README.md`
- Update with new location of `crypto/README.md` file

`docs/workload-tutorial/README.md`
- Fix incorrect links to `stage_2` source files
- Use proper BNF notation, `<>`, for `name`, `key`, and `count`

`docs/workload-tutorial/*`
- Merge stage 3 with stage 1 and 2 code
- Remove differences between templates and `stage_{1,2,3}` code
- Make copyright dates consistent between original and modified files

`common/.../file_io_wrapper.h`
- Fix typo
