+++ docs/workload-tutorial/hello_world/stage_1/logic.h
@@ -1,4 +1,4 @@
-/* Copyright 2019 Intel Corporation
+/* Copyright 2020 Intel Corporation
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
@@ -18,4 +18,3 @@
 #include <string>
 
 // Add actual workload definitions here ...
-
