+++ docs/workload-tutorial/hello_world/stage_1/plug-in.h
@@ -1,4 +1,4 @@
-/* Copyright 2019 Intel Corporation
+/* Copyright 2020 Intel Corporation
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
@@ -54,4 +54,3 @@ public:
 }  // namespace HelloWorld
 
 using namespace HelloWorld;
-
