+++ docs/workload-tutorial/hello_world/stage_3/CMakeLists.txt
