+++ docs/workload-tutorial/hello_world/stage_3/logic.h
