+++ docs/workload-tutorial/templates/CMakeLists.txt
@@ -1,4 +1,4 @@
-# Copyright 2018 Intel Corporation
+# Copyright 2020 Intel Corporation
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,9 +13,6 @@
 # limitations under the License.
 
 CMAKE_MINIMUM_REQUIRED(VERSION 3.2 FATAL_ERROR)
-
-# Common components for both trusted and untrusted common libraries
-################################################################################
 
 FILE(GLOB PROJECT_HEADERS "*.h")
 FILE(GLOB PROJECT_SOURCES "*.cpp")
