+++ docs/workload-tutorial/templates/plug-in.h
@@ -1,4 +1,4 @@
-/* Copyright 2019 Intel Corporation
+/* Copyright 2020 Intel Corporation
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
@@ -56,5 +56,3 @@ public:
 
 // replace $NameSpace$ below with a workload namespace, e.g. HelloWorld
 using namespace $NameSpace$;
-
-
